fix(phase3): remove unnecessary type conversions and prealloc slices (#21)
Phase 3a - Quick Wins: Fix unconvert (8) and prealloc (16) linter issues

UNCONVERT (8 issues) - Remove unnecessary type conversions:
- internal/auth/secure_store.go: Remove int() around syscall.Stdin
- internal/dao/slurm_adapter.go: Remove int() conversions for CPUs
- internal/mock/validator.go: Remove int() around syscall.Stdin
- internal/setup/wizard.go: Remove int() around syscall.Stdin (2x)
- internal/ssh/key_manager.go: Remove string() conversion
- plugins/observability/views/observability.go: Remove uint64() conversion

PREALLOC (16 issues) - Preallocate slices with known capacity:
- internal/auth/load_balancer.go: Preallocate with len(endpoints)
- internal/config/templates.go: Preallocate with len(templates) and len(categories)
- internal/discovery/auto_discovery.go: Preallocate with capacity 3
- internal/discovery/cluster_discovery.go: Preallocate with len(clusterMap)
- internal/plugin/manager.go: Preallocate with len(plugins)
- internal/plugins/manager.go: Preallocate for commands, views, bindings
- internal/ssh/agent_auth.go: Preallocate with len(signers) (2x)
- internal/streaming/eventbus.go: Preallocate with len(subscribers)
- internal/ui/components/multi_select_table.go: Preallocate with len(selectedRows)
- internal/ui/widgets/config_manager.go: Preallocate with len(arr) and len(parts)
- internal/views/enhanced_terminal_view.go: Add capacity for nodeOptions

These are performance optimizations that reduce allocations and improve clarity.

Total issues fixed: 24 (unconvert: 8, prealloc: 16)
Remaining Phase 3: unparam (37), noctx (23), gocritic (51), unused (1)

+++ internal/dao/slurm_adapter.go
@@ -630,7 +630,7 @@ func convertNode(node *slurm.Node) *Node {
 		case "idle":
 			allocCPUs = 0
 		case "mixed", "allocated":
-			allocCPUs = int(node.CPUs / 2) // Estimate 50% utilization
+			allocCPUs = node.CPUs / 2 // Estimate 50% utilization
 		case "down", "drain":
 			allocCPUs = 0
 		default:
@@ -660,7 +660,7 @@ func convertNode(node *slurm.Node) *Node {
 	memoryTotalMB := int64(node.Memory)
 
 	// Calculate idle CPUs (total - allocated)
-	idleCPUs := int(node.CPUs) - allocCPUs
+	idleCPUs := node.CPUs - allocCPUs
 	if idleCPUs < 0 {
 		idleCPUs = 0
 	}
